refactor(platform): split runtime platform module
Refs #290 task 13.
