+++ src/crosshook-native/crates/crosshook-core/src/platform/tests/mod.rs
@@ -1,0 +1,5 @@
+mod common;
+mod detect;
+mod gateway;
+mod steam_deck;
+mod xdg;
